requirements.txt
streamlit
streamlit-chat
langchain
langchain-community
sentence-transformers
faiss-cpu
PyPDF2
pdfminer.six
transformers
torch
huggingface-hub
python-dotenv
pypdf
langchain_openai
chromadb
streamlit>=1.22.0
streamlit-chat>=0.1.1
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
azure-identity>=1.13.0
pypdf>=3.15.1
chromadb>=0.4.13
tiktoken>=0.5.1